Refactor files a bit to organize separately
